+++ CHANGELOG.rst
@@ -39,6 +39,7 @@ Config Updates
 
 - Added Correlated noise type as a generalization of the more specific COSMOS noise type. (#731)
 - Added galaxy_sample input type with corresponding SampleGalaxy GSObject type. (#731)
+- Added COSMOSValue and SampleValue value types. (#954)
 - Changed the way the internal random number sequence works so that running multiple simulations
   with sequential random seed values doesn't end up with duplicated random values across the
   two (or more) simulations. (#1169)